Skipping tests that make repeated connections on OSX
Stem's tests currently hang on Mac OSX, and that doesn't look to be getting
better any time soon...
https://trac.torproject.org/5917

I've narrowd this down to being either a python, tor, or mac bug with how
repeated socket connection/closures are handled on that platform. For now
skipping the tests that encounter this issue if we're running on a mac.