Supporting multiple extra-info transport lines
The 'transport' lines being added to extra-info descriptors can appear any
number of times, so storing them as a list. These options are weird in that
they have an "address:port [arglist]" which never appears in the wild since
they both only appear on bridge descriptors and are scrubbed from bridge
descripotrs.