First draft of torexport, 3 functions defined to handle different use cases.
First draft of torexport, 3 functions defined to handle different use cases.
csv_file_exp() takes a list of descriptors and exports their attributes as a
csv file. descriptors_csv_exp() takes a list of descriptors and returns a
generator.  Each iteration, this function returns a line of csv representing
the attributes of a single descriptor object.  descriptor_csv_exp() returns
a single line of csv based on the attributes of a sole descriptor object.