Calling QUIT prior to disconnecting socket
Making the Controller's close() method try to be a nice controller by issuing a
QUIT call before detaching the socket. This is a best-effort call (errors
ignored) since the caller just wants the socket shut down and we don't need the
QUIT to succeed to do that. Original change by gsathya and discussed on...
https://trac.torproject.org/6234