Moving descriptor contructor to mocking module
We'll need to construct server descriptors in both their unit tests and when
testing the exporter. The mocking module is both for function mocking and
creating mock objects so this is the perfect place for it. I might move the
extrainfo descriptor constructer here later if it seems useful.

Personally I'm finding the current exporter tests and mocking changes very
confusing so reverting the mocking changes and writing some simpler tests.