Adding an easy method for getting a controller instance
Adding a 'prompt' script that kicks off a python interpretor with a readily
available stem.control.Controller instance. This starts a tor test instance if
it isn't already running and offers to shut it down when done.

The purpose of this script is to make it easy to get a controller instance we
can exercise. The sh shebang makes this *nix specific, though I suppose windows
users can still use it by running 'bash prompt'.

Example usage:

  atagar@morrigan:~/Desktop/stem$ ./prompt
  Welcome to stem's testing prompt. You currently have a controller available
  via the 'control' variable.

  >>> control.get_info('version')
  '0.2.1.30'
  >>> quit()

  Would you like to stop the tor instance we made? (y/n, default: n): y