Raising TypeError when metrics header is unrecognized
When we get a metrics header for a format we don't recognize the
_parse_metrics_file() function is supposed to throw a TypeError (and says it
does in a comment). However, it didn't.

This caused unrecognized metrics types to be parsed like a valid descriptor
file containing zero descriptors. Fixing and adding a test for this use case.

Also fixing a couple other minor bugs I ran into while fixing this:

* Some tests had an assertTrue() rather than assertEqual(), causing the
  assertions to almost always evaluate to True.

* The UnrecognizedType we threw for text files didn't have its mime type set.