Inverting network status' document and router relationship
This code was trying very hard to produce a network status document that
contained router entries. This is right and proper from an object oriented
standpoint, but not from a practical one. The document cannot contain the
router entries unless we keep them all in memory, and for the vast majority of
users that is not desirable.

In later revisions Ravi addressed the memory concerns by turning the
router_descriptors attribute into an iterator. This works, but it's confusing
and the router_descriptors' itertor only works as long as we keep the file open
(or buffer the content).

Instead, inverting the relationship so this module provides an iterator of
routers and those routers have a reference to the document they came from. As a
side benefit this lets us tidy up the code a bit.

There like are some use cases where we would prefer a document object that
contains routers, so I'll probably add that option too later.

This also gets rid of the "Flavour" enum. I'm not entirely sure what it was for
- it wasn't documented and its NS value was unused. Maybe it was vestigial from
some prior changes.