Hybrid approach between iterator and attribute
Ok, changed my mind. At least partly. If the user is using parse_file() then we
want to provide them with router entries and a bare document, like I was.
However, if they're calling the document constructors then they've already read
the complete document into memory so there's little reason not to parse it.

In short, documents now have a 'routers' attribute. That attribute is filled
with the parsed routers if we use the constructor directly, and left empty if
we used parse_file().