Error due to KEYWORD_LINE regex change
The KEYWORD_LINE regex fix a few commits back broke one of the server
descriptor unit tests because it caused lines with only a keyword (ex.
"platform") to have a None value. Before the regex change they had an empty
string value.

I'm not really sure which is proper, but for now keeping the prior behavior.