Dropping unknown_flags attribute
I suggested an unknown_flags attribute that would have flags which stem didn't
recognize but, on reflection, this wasn't a very good idea. Splitting the
attributes mean that users have multiple places to look for the relay's flags.

We were also failing validation if there were any unknown flags and that is
definitely wrong. New flags are perfectly valid and will happen in the future -
those new flags should not break stem's parser. This was probably confusion
from when I asked for the cached-consensus test to check for new flags. This is
something separate and added it with this change.