Rewriting RouterStatusEntry parser
The networkstatus module has spotty error checking, for instance...

* the unrecognized_lines attribute is never initialized, so anything touching
  it will trigger a NameError

* missing values on an 'r' line will result in an IndexError, and extra values
  are ignored

* malformed nicknames, ip addresses, and ports go undetected

* non-integer ports or bandwidth/measured values give an 'invalid literal'
  error, which isn't terribly helpful

* extra entries on a 'w' line inappropriately cause an error

* malformed exit policies can cause an error when validation is disabled

* we error if there's new keywords (the spec doesn't seem to say they're
  disallowed...)

From the look of this code the 'validate' flag seems to have been
misinterpreted as meaning "raise an exception if we run into something
unexpected, like a new keyword or flag". This is wrong. Stem should *never*
error when processing spec conformant content. Rather, the 'validate' flag
means "raise an exception when we parse something that violates the spec". When
that flag is False the parser should *never* raise an exception, making a
best-effort attempt to parse even malformed content.

I'm a bit sad to see that there are absolutely no unit tests for this module.
That was the bulk of the testing for the server_descriptor and
extrainfo_descriptor modules. Guess I'll write that next.