Checking for mandatory and disallowed fields
There's several restrictions on a valid network status document, some of which
are which fields it does and does not contain. Validating that manitory fields
apprear, and that fields which shouldn't appear don't.

This also drops my _get_entries() helper function in favor of
_get_descriptor_components() since we're now dealing with documents which can
contain key blocks. There wasn't really enough of an advantage to
_get_entries() to justify the duplicate code anyway.

Also fixing the _get_descriptor_components() I wrote so it return footer
content in the header when the content doens't contain any 'r' lines.

This currently keeps the old parser (renamed to "_parse_old()") since I haven't
addressed any of the actual parsing yet.