Unit test for misordered fields
To be valid a network status document's fiends need to appear in a particular
order. I'm about to add vaidation for this, but the parser is being finicky.
I'm tired of trying to fix incrementally fix it, so skipping this test for now
and moving on to the parser rewrite.