Validating that fields aren't duplicated
Almost all of the fields in a network status document can only appear once.
Checking that this is a case. This also includes a unit test for the
validation. I'm a little surprised that the old parser doesn't choke on
duplicated content, but I'm about to replace it so probably not worth
investigating.