Breaking up the header and footer from NetworkStatusDocument
The NetworkStatusDocument class was starting to get monsterous, and it was just
gonna get worse. A network status document consists of four sections...

- header
- authorities
- router status entries
- footer

Making the NetworkStatusDocument a thin container for these four, and making
separate classes for them. This has made the code much nicer.

The only disadvantage that I've seen is that validation is done in pieces so
if, for instance, you're missing mandatory fields from both the header and
footer you now won't be told about both in a single error message. Instead
the header will be parsed first, fail, and just tell you about those.

That said, this is a pretty minor regression and well worth the improved
maintainability.