Generalizing how router entries and authorities are parsed
The _get_routers() and _get_authorities() were essentially doing the same
thing. Replacing both with a more general _get_entries() helper that reads a
range of the document and constructs instances for it.

Taking advantage of this nicer helper's keyword arguments to make the code more
readable (functions that take a ton of positional args are is a pita).

I'm a bit surprised (and concerned) at how easily this passed unit tests. No
doubt I've broken the integ tests but I'm not putting any effort there until
I've finished the document parser rewrite.