Unit tests for parse_file() and router entries
On reflection the reason that the prior changes passed the unit tests so easily
was because the parse_file() function and inclusion of router status entries
was completely untested by my unit tests. The RouterStatusEntry class itself it
tested, but not its inclusion in a document. Integ tests would certainly cover
this, but I want the unit tests to exercise everything too.

Adding the missing tests and some fixes for issues that they revealed.