Generalizing mocking.get_server_descriptor()
All of the descriptor unit tests (server, extrainfo, and document components)
need a similar type of mock data. I've been doing this in an adhoc fashion but
this can stand for a bit of uniformity. Starting off by cleaning up the server
descriptor mocking.

As per the convention for the mocking module, making it provide a mock object
by default with an option to get the raw content.