Supporting microdescriptor flavored consensuses
Adding support for microdescriptor flavored consensuses into the
NetworkStatusDocument class. It made sense to have a separate class for it, but
on the other hand it *is* still a v3 consensus and the only impact the flavor
has is alternate router status entries so just blending a 'flavor' and
'is_microdescriptor' attribute in.