Merging testrc.sample with settings.cfg
We run the tests all the time but never with a configuration file so
deemphasising the capability by merging the sample rc file with the
settings.cfg (which is actually... well, used). If someone really wants to
configure the tests this way then they might as well look at the settings.cfg
anyway since it'll provide the full range of options.