Moving tutorial tests to be unit tests
Both of our integ tests for the tutorial examples were a bit clunky, the
controller test because it could only run with newish tor versions and the
descriptor test because it had to make use of metrics descriptors. Oh, and they
didn't actually assert anything.

Moving the tests to be unit tests instead. This included greatly expanding our
mocking capabilities, supporting arbitrary mock objects. I'm not sure if the
end result is any better than our prior integ tests since we're mocking pretty
much everything (and in the case of the descriptor test it's pretty ugly). On
the other hand the tutorial tests are primarily for basic syntax and
compatability with our current version of stem so guess it's ok...

All this said, I *really* like our arbitrary object mocking capability...