Only including prepared API docs
Our sphinx-apidoc made it so all our pydocs were included in our built
documentation. This is great, except that those documentation dumps are
completely unreadable. Dropping the sphinx-apidoc call in favor of including
specific autodoc pages we've prepared.

Made some revisions to stem.version's documentation and including that to
start.