Line wrapping pydoc clauses
I had been avoiding line wraps in :var:, :param:, and :return: entries to avoid
confusing the spinx compiler but on reflection it's output is uneffected when
we indent the following line.

Just wrapping the lines that were really unweildy rather than going for a
strict eighty character width. Sorry
lines-of-code-must-be-shorter-than-my-iphone-screen zelots.