Revised API docs for stem.response
This is a little different from the other modules in that we're including
stem.response.__init__ as an automodule then autoclasses for all of the
contents. The response classes each have precious little documentation to
having them each on their own page would be pointless.

Also moving the AuthMethod enum to the connection so we can drop the
protocolinfo module documentation (it wouldn't be included since we're using
autoclass).