Supporting files in whitespace checks
We called os.walk() when determining the files for which we want to check
whitespace. However, when presented with a file rather than a directory this
causes us to not check anything...

>>> list(os.walk("/tmp/foo"))
[]

This broke our attempts to check 'run_tests.py', and let a couple whitespace
issues slip in. Fixing get_issues()'s handling for individual files.

Issue caught by Eoin on...
https://trac.torproject.org/7263