Image credits and licenses
Noting where the site images come from. Almost all of them are from the
collection I've gathered of free image resources
(http://www.atagar.com/freeMedia/).

I've tried pretty hard to avoid images with unknown licenses but there is one
exception: the anvil schematic for the API link. This is a fantastic image
and I've searched quite a bit to figure out where it originally came from
(found it on a blog but doubt that's the source). Pity...