stem.process integ test with RUN_SOCKET target broken
Ok, this puzzled me for a while. Our stem.process integ tests make
'/tmp/stem_integ' to be its temporary data directory for the run, then cleans
it up afterward.

This is great, except that the RUN_SOCKET causes us to run with the following
in its torrc...

ControlSocket /tmp/stem_integ/socket

It's a coincidence that both use the same tmp directory, but it's not
necessarily bad. What was bad, though, is that...

* the stem.process tests broke since the mkdir call for stem_integ failed
* if that hadn't failed our tearDown would have deleted stem_integ, and with it
  our control socket causing all further tests to fail

We should probably use the tempfile module to make our stem.process data
directory, but for now just being lazy and having it account for the RUN_SOCKET
target.

Caught by Ravi on...
https://trac.torproject.org/7284