Adding event handling support to the Controller
My initial plan was to add event listener support similar to TorCtl but with
automated handling of the SETEVENT calls. While working on this meejah
suggested accepting functors instead (like txtorcon). On reflection this is
both much nicer for our callers and easier for us to support.

Adding functions for adding and removing listeners, with a simple integ test.
Next up will be the Event classes...