Unit test for add_event_listener() example
Adding a unit test for the pydoc example in add_event_listener(). I really,
really don't like how the tests for these documentation examples are turning
out. The end result looks precious little like the example and is convoluted as
hell.

We should revisit how we do documentation example testing in the future, but
that'll be a subproject in itself.