Reverting comment change about tor error messages
The prior spelling fixes condensed a comment that we had about the error
message tor will provide when we use the wrong controller password. The change
makes sense, but these lines are here to tell us the exact error message tor
provides (after which it's hopefully reasonably obvious why we check for the
"Password did not match HashedControlPassword" part of it).