Support for ADDRMAP events
There's a special spot in hell for whoever decided to allow for quoted values
in events. This implements and adds testing for ADDRMAP events. Unlike TorCtl
we aren't falling back on a regex for the... er, 'wonderful' quoted stuff, but
rather including quoted value support in the Event parser.

Got test data by visiting a few sites in TBB...

650 ADDRMAP check.torproject.org 38.229.72.22 "2012-11-18 22:48:34" EXPIRES="2012-11-19 06:48:34"
650 ADDRMAP ocsp.digicert.com 5.63.145.124 "2012-11-18 21:53:42" EXPIRES="2012-11-19 05:53:42"
650 ADDRMAP www.atagar.com 75.119.206.243 "2012-11-19 00:50:13" EXPIRES="2012-11-19 08:50:13"