Testing for STATUS_* events
I had grand plans to have a StatusEvent subclass for each of the causes.
However, on reflection that would encompass a bundle of work that's almost as
large as all of the other events put together. It might still be a neat thing
to do someday, but there's plenty of higher priority things to do first.

Adding tests for all of the event examples that I collected earlier. This
includes a bit of a hack to make quoted key/value mappings work. Previously
we've only parsed quoted mappings if they belonged to the _QUOTED listing, but
STATUS_* events sprinkle them all around so for forward compatability we can't
simply enumerate them.