Simplify check for if event attributes are unrecognized
Whenever event parsing handles an attribute for which we have an enum it checks
to see if the value is recognized by stem. If not then it logs at the INFO
level.

Replacing these boilerplate checks with a helper function. I'm also adding a
test to check that we're actually logging these messages.