Purging use of the variable 'i'
A pox upon whoever taught me to use variables like i, j, and k for loop
constants. It's ungrepable and a practice that should be purged via fire and
pitchforks.

I'm keeping it in for list interpolation (since it's localized to a single
line), but replacing it with 'index' everywhere else. I'm also replacing a few
points that do...

for i in range(len(my_list)):
  my_item = my_list[i]

... with calls to enumerate().