Reducing ExitPolicyRule's memory requirements
Frequently there's several ExitPolicyRule entries per an exit policy, so when
we pull the consensus we get quite a few instances of this class. Making the
following changes to reduce the memory requirements...

* Dropping the rule attribute. The string representation of the rule should be
  good enough, if not better for callers.

* Replacing the address_type attribute with a method for getting it. This lets
  us store the address type as an integer within our class.

* Replacing the mask attribute with a method for getting it. The ip mask
  representation is very rarely useful, so there's little reason to store it
  unless it's requested.

* Lazily loading the integer reprentation of our address and mask, both
  speeding up our constructor and avoiding it entirely if our caller never uses
  is_match()

This lowers the memory requirement for loading the full consensus on my netbook
from 5.5% to 5.1% (a 7% drop). This is a drop in the bucket compared to the
prior commit, but between the faster constructor runtime and squeezing out a
little more performance it's still worth it.