Use threading.Event in event listener tests
wait() up to two seconds, for a threading.Event to be set() in the
test event listener before continuing assertions.

This should speed up success or failure in these tests, even if not by
much.  test_event_handling() goes from a constant time of 5 seconds, to
a variable 4-8 seconds, with 4 seconds most common in my testing.
test_reattaching_listeners() goes from a constant 6 seconds, to a variable
4-6 seconds.

Signed-off-by: Sean Robinson <seankrobinson@gmail.com>