Revising error handling for get_conf()
The Controller's get_conf() function was buggy if the configuration value was
unset. It's documented as returning the default, but instead it returned None
or [None] depending on if 'multiple' was set.

On reflection the pydocs for these functions were pretty unfriendly, so giving
them an overhaul. Functionally this changes two things...

1. When you query for an unset value it will...

   * provide the default value if you gave it one
   * provide None if multiple was False, and [] if it was True

2. The get_conf_map()'s 'default' argument is now the defaultly *mapped value*
   rather than the default *response*. For instance...

   # old behavior
   get_conf_map('NoOption', 'Blarg') => 'Blarg'

   # new behavior
   get_conf_map('NoOption', 'Blarg') => {'NoOption': 'Blarg'}

   This means that the get_conf_map() pydocs can now assure the caller that
   they'll always get a dict, and it also better matches the get_conf()
   handling for default.