Fixing eronious 'missing with' warning
Running our tests gave a working that network.py was missing an import of the
'with' statement. This was wrong - the with statement was in its pydocs. The
checker tries to be smart about what is and isn't documentation by looking for
block quotes but an unusual string confused it...

  """ Base error for proxy issues. """

I avoid using block quotes unless things span multiple lines which is why I
didn't account for this. We could fix the checker (... which isn't a bad idea)
but for the moment I'm just dropping the odd block quotes since it'll probably
introduce extra whitespace before and after the message in the sphinx docs.