Uniformly providing a default argument for getters
As disussed on 'https://trac.torproject.org/7832' it's important from a
usability standpoint for all of our getter methods to support a 'default'
argument. It's also nice to have class-wide uniformity on this since it makes
the Controller class more intuitive to use.

This also reorganizes the method listing in the header to group related
methods, and has a couple small bits of backward incompatability (trying to get
these breakages in early before making our initial release announcement)...

* renamed protocolinfo() to get_protocolinfo() to match the other getters
* renamed one of the 'circuit' arguments to 'circuit_id' (we did that for all
  of the methods but one)