Network status integ tests should check for unrecognized lines
Probably the most useful integration test that we have is the check for
unrecognized lines in the server descriptors. That one test has found dozens of
upstream tor bugs (mostly new additions without spec updates).

It turns out that we were missing this check for network statuses, and it's
already tripping on an issue. Oops.