Conforming to E126 (continuation line over-indented for hanging indent)
Ok, I haven't a clue what the hell this one was about. In one spot pep8
complained about the hanging indent. However, it looks right to me. For some
odd reason it wants the hanging indent to match the
'stem.descriptor.networkstatus' import above - maybe this is a bug with pep8.

Oh well. Changing our code to shut it up.