Skipping E111 and E121 (four space indentations)
I strongly prefer two space indentations. It reads perfectly well and leads to
shorter lines. I've never heard a coherent argument why four spaces, eight
spaces, tabs, or any of the other measurement is actually superior.

When it comes to this bike shed I'm keeping the color as it is for now.