Support process renaming
Renaming our python process from "python our_script.py" to something more
sensible is a very handy capability to have. Thanks to Jake arm has had this
for years, which makes it easier to identify in top and ps listings.

We'll want this capability for other controllers so snagging arm's utility and
cleaning it up.