Adding a BaseController is_authenticated() method
I've been running into some situations where I need to know if the controller
is not only connected, but authenticated (since that determines if it's...
well, usable).

A use case where this could get confused is...

* User gets a ControlSocket instance.
* Manually authenticates via the send() and recv() methods.
* Makes a new Controller object from this control socket.

Unfortunately this is *exactly* what our test runner did, so tweaking it so
that it hits our post_authentication method (I'm surprised that it hadn't
bitten us before).