Letting add_event_listener() work when unauthenticated
If we called add_event_listener() prior to authentication then we'd throw an
error when calling get_version() to figure out if we meet the requirements.

Changing the behavior so that we check requirements and issue SETEVENTS if
able, but if not then simply enquing the listener. It'll then be attached
during our post-authentication hook.

The change that I'm more interested in this though is that our
post-authentication hook is less likely to go belly up. If we, say, added a
listener then attached the Controller to an older tor instance then our
SETEVENTS call could fail. This was an all-or-nothing call. :(

Changed it so that the hook will re-attach the events that it can, then warn
about the rest.