Always processing @type headers
The parse_file() function only consumed the @type annotation when we didn't
have a descriptor_type nor recognized filename. Changing it so we always
consume the @type header.

I'm also changing the priority order of parse_file() so if the file could be
both a metrics archive or cached descriptor (for instance a file named
'cached-consensus' that starts with a @type annotation) then we process it as a
metrics archive.