Dropping hardcoded python 2.x check
Our prereq module asserted that stem isn't compatable beyond the 2.x series.
Clearly I'm trying to 'fix' that with this branch.

Ideally we'd keep this check for our python 2.x codebase, and strip it when
we do a 2to3 conversion. However, that's more error prone than I'd like so
simply dropping the check entirely.