Converted _required_fields() encounter TypeErrors
The server and extrainfo descriptors' _required_fields() functions get
converted by 2to3 into a line that tries to concatenate a list with a tuple
(raising a TypeError). On reflection the converted version which uses list
comprehension is a little nicer than filter(), so using a variant of that which
makes everyone happy.