Returning unicode from stem.util.system.call()
The call() method is the basic call for shelling out and, like tor's stdout
earlier, its output was ASCII bytes. This is fine in python 2.x, but causes
problems for python 3.x. Both are fine with this being unicode output so going
with that.

======================================================================
ERROR: test_get_cwd
----------------------------------------------------------------------
Traceback:
  File "/home/atagar/Desktop/stem/test/data/python3/test/integ/util/system.py", line 334, in test_get_cwd
    self.assertEquals(None, stem.util.system.get_cwd(99999))
  File "/home/atagar/Desktop/stem/test/data/python3/stem/util/system.py", line 528, in get_cwd
    elif results[0].endswith("No such process"):
TypeError: expected an object with the buffer interface