Providing alternative methods for parsing a NetworkStatusDocument
Adding support in both the DescriptorReader and parse_file() function for three
ways of parsing network status documents...

a. Provide the router status entries (ie. the current behavior).

b. Provide the document itself with the router status entries that it contains.
   This has the biggest cost in terms of upfront parsing time and memory usage,
   but provides the caller with everything they might want.

c. Provide the document but skip reading the router status entries. A handy
   option of you just care about the document's header/footer.

Now that we have these capability I'm further simplifying the descriptor API a
bit. The network status docs encouraged users to use the NetworkStatusDocument
constructors to achieve option 'b' above, but now that it's in the reader and
parse_file() there's no reason for them to do that.

Users should now *always* use either the DescriptorReader or parse_file(). If
they don't then they're off the reservation.