Using port lists for addresses_v6 rather than ranges
Huh, I wonder where I got the idea that 'a' lines had port ranges. Dropping
that. According to the spec the 'a' lines should be parsed in the same way as
'or-address'. However, I suspect that the spec is a little off here - checking
if it can contain IPv4 addresses...

Caught by Karsten on...

https://trac.torproject.org/8036