Dropping the 'path' argument from stem.descriptor.parse_file()
File objects have a 'name' attribute that we can use to guess the path. This
isn't entirely reliable, but nothing is...

http://stackoverflow.com/questions/2458676/absolute-path-of-a-file-object

The path argument was only there to support the descriptor reader. Now that
parse_file() is something for our users it's nice to get rid of arguments they
can't use.