Don't pass unsupplied optional args to wrapped file
Karsten reports that the descriptor reader is failing for him with...

  ParsingFailure: readlines() takes exactly 1 argument (2 given)

This is because to add python 3.x support we had to wrap the parse_file()'s
file argument with a wrapper. In most cases this wrapper is a passthrough, but
it converts the read() and readlines() output to unicode.

I based the methods of the wrapper on a StringIO file object, but evidently
other files (most likely the tarfile) deviate a bit. Addressing the readlines()
instance and one other that might cause troubles.