Mocking didn't properly revert static methods
When working on our tutorials I expanded our mocking module to handle static
methods. This worked well for our unit tests, but when we reverted the mocks we
did so as standard rather than static methods.

This in turn broke all of our integ tests that relied upon
Controller.from_port(). This went undetected because it only manifests when
*both* unit and integ tests are run together (and I've only been running one or
the other recently).