stem.process.launch_tor() doesn't work without stdout logging
The launch_tor() function uses tor's stdout logs to figure out when it has
finished bootstrapping. This, of course, doesn't work if tor isn't logging to
stdout...

https://trac.torproject.org/7677

Changing launch_tor_with_config() to slip in a 'Log NOTICE stdout' option so it
always works, and adding a note to launch_tor() warning users of the
requirement.