Dropping 'last updated' timestamp from site
While I'd kinda like to have this on our site, it's stupidly broken. This is
probably due to sphinx caching (it only makes static files for things that have
changed). Presently it says that our site was last updated on December 31st
which is... very wrong. No date is better than the wrong date so dropping it.

https://trac.torproject.org/7780