Dropping the memory usage assertion when reading descriptors
Our integ test for reading descriptors tries to check that we aren't consuming
gobs of memory. Unfortunately this assertion seems to be platform dependent (it
fails on OSX) so dropping the assertion. Yea, yea, I should be fixing it but we
haven't had this issue for a very long while and it isn't likely to reoccure.