Storing raw descriptor content as bytes
Tor descriptors are byte strings. While most content is ASCII (and hence maps
to UTF8) there are a couple exceptions: the server descriptor's contact and
platform lines which are arbitrary byte content. Usually this is unicode but on
occasion it isn't (14 of 3044 in my old consensus cache).

In those cases we converted the content to unicode and replaced the
un-decodeable characters. This in turn caused our _verify_digest() check to
report the descriptor content as being malformed if the user had pycrypto
installed.

Changing the raw contents the Descriptor class tracks to be bytes instead of
unicode. This actually leads to far more elegant code, letting us drop the
UnicodeReader which was always a pretty gross hack.

This breaks python 2.5 compatability by using the io.BytesIO class. Presently
I'm checking with tor-dev@ to see if anyone objects with dropping python 2.5
support entirely...

https://lists.torproject.org/pipermail/tor-dev/2013-March/004551.html

If the answer is...

* yes, drop support: we'll replace test.mocking.BytesBuffer with io.BytesIO
  (they both do the exact same thing)

* no, keep support: we'll need to move test.mocking.BytesBuffer to a util and
  use it instead of io.BytesIO