Changing isinstance() checks to look for both bytes and unicode
Commit b12e9c0 broke a couple of our integration tests because it caused us to
call controller methods with unicode rather than str arguments...

https://trac.torproject.org/8576

This should be fine - python 2.x can usually use bytes and unicode inputs
interchangeably. The problem was that we routinely use isinstance() checks to
see if an input is a string and, if it isn't, assume it's a collection. As a
result we were calling...

GETINFO m d / n a m e / f o o

... rather than...

GETINFO md/name/foo

Tor in turn errored saying that those single character getinfo options didn't
exist. To fix this I'm changing most of our isinstance() checks to look for
both bytes and unicode.