Having ExitPolicyRules report /0 as a wildcard
ExitPolicyRules can apply to both IPv4 and IPv6. Because of this I treated * as
a wildcard, but *not* /0 (since 0.0.0.0/0, though it applies to all IPv4
addresses, doesn't cover IPv6). However, in most cases (everywhere except the
torrc?) Tor's exit policies are limited to IPv4. As such, /0 should usually be
treated as a wildcard. Change suggested by Aaron Johnson.