Simplifying use of digest values for microdescriptors
To be useful a microdescriptor often needs to be matched against its
corresponding router status entry. The design of microdescriptors makes this
suck - the only way to do so is to iterate over all router status entries
looking for one with the corresponding digest.

Adding examples for doing this and making the following changes to make the
matching easier...

* changing the digest value of RouterStatusEntryMicroV3 from base64 to hex
* adding an attribute with the hex digest to Microdescriptors